pyproject.toml
[tool.poetry]
name = "daynight-theme"
version = "0.1.0"
description = "Set automatic Gnome theme during day and night"
authors = ["Mikedev <mik3dev@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.7,<4.0.0"
path = ">=15.0"
PyYAML = ">=5.0"
rich = ">=10.0.0"
tenacity = ">=7.0.1"
dateutils = "^0.6.12"

[tool.poetry.dev-dependencies]
pytest = ">=6.0"

[tool.poetry.scripts]
daynight-theme = 'daynight_theme:main'
daynight-theme-config = 'daynight_theme.daynight_theme_config:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
